+++ docs/pages/documentation/general/_General.vue
@@ -3,8 +3,8 @@
         <router-link to="layout" class="nav-item is-tab">Layout</router-link>
         <router-link to="icon" class="nav-item is-tab">Icon</router-link>
         <router-link to="dropdown" class="nav-item is-tab">Dropdown</router-link>
-        <router-link to="tooltip" class="nav-item is-tab">Tooltip</router-link>
         <router-link to="modal" class="nav-item is-tab">Modal</router-link>
         <router-link to="tabs" class="nav-item is-tab">Tabs</router-link>
+        <router-link to="tooltip" class="nav-item is-tab">Tooltip</router-link>
     </div>
 </template>