+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.4.5
 
+* Update modal docs example to not use the ``component`` prop, works better for events and is more common — Old way is not deprecated
 * Remove programmatic component modal from docs since isn't a good practice and was causing too many issues
 * Fix autocomplete ``focus()`` method
 * #106 Fix prompt dialog vue-warning